+++ CodinGame/CodeRoyale/CodeOfKutulu.rs
@@ -43,13 +43,23 @@ impl Game {
         }
     }
 
+    fn get_closest_wanderer(&self, player: &Explorer) -> Option<i32> {
+        if let Some(w) = self.wanderers.iter().map(|(&i, u)| {
+            (i, manhattan_distance(player.get_coord(), u.get_coord()))
+        }).min_by(|a, b| a.1.partial_cmp(&b.1).unwrap()) {
+            Some(w.0)
+        } else {
+            None
+        }
+    }
+
     fn handle_explorer(&mut self, player_id: i32) {
         let player = self.explorers.get(&player_id).unwrap().clone();
         self.explorers.remove(&player_id);
 
         let mut move_coord = player.get_coord();
 
-        if let Some(i) = get_closest_wanderer(&self.wanderers, &player) {
+        if let Some(i) = self.get_closest_wanderer(&player) {
             let wanderer_c = self.wanderers.get(&i).unwrap().get_coord();
             let player_c = player.get_coord();
 
@@ -241,17 +251,6 @@ fn manhattan_distance(a: (i32, i32), b: (i32, i32)) -> i32 {
     (b.0 - a.0).abs() + (b.1 - a.1).abs()
 }
 
-fn get_closest_wanderer(wanderers: &HashMap<i32, Wanderer>,
-                        player: &Explorer) -> Option<i32> {
-    if let Some(w) = wanderers.iter().map(|(&i, u)| {
-        (i, manhattan_distance(player.get_coord(), u.get_coord()))
-    }).min_by(|a, b| a.1.partial_cmp(&b.1).unwrap()) {
-        Some(w.0)
-    } else {
-        None
-    }
-}
-
 #[derive(Debug, PartialEq)]
 enum Direction {
     LEFT,
@@ -260,6 +259,7 @@ enum Direction {
     DOWN
 }
 
+// Should take an Explorer
 fn get_possible_moves(map: &Vec<Vec<char>>, coord: (i32, i32))
                       -> Vec<Direction> {
     let mut result = vec![];